requirements.txt
langgraph>=0.0.11
langchain>=0.0.350
langchain-core>=0.1.10
langchain-openai>=0.0.2
langchain-groq>=0.0.1
langchain-community>=0.0.10
langchain-experimental>=0.0.47

# LLM APIs
openai>=1.3.0
groq>=0.4.2

streamlit
# Environment variable management
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
pydantic>=2.4.2

# Search tools
duckduckgo-search>=3.9.0

# Jupyter environment
jupyter>=1.0.0
ipython>=8.15.0

# Data processing
pandas>=2.1.0

# Type hints
typing-extensions>=4.8.0

# Visualization and graphs
graphviz>=0.20.1